+++ app/jobs/area_visits_calculation_scheduling_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AreaVisitsCalculationSchedulingJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    User.find_each { AreaVisitsCalculatingJob.perform_later(_1.id) }
+  end
+end