+++ db/migrate/20240721183005_create_visits.rb
@@ -5,6 +5,11 @@ class CreateVisits < ActiveRecord::Migration[7.1]
     create_table :visits do |t|
       t.references :area, null: false, foreign_key: true
       t.references :user, null: false, foreign_key: true
+      t.datetime :started_at, null: false
+      t.datetime :ended_at, null: false
+      t.integer :duration, null: false
+      t.string :name, null: false
+      t.integer :status, null: false, default: 0
 
       t.timestamps
     end