+++ db/schema.rb
@@ -168,6 +168,11 @@ ActiveRecord::Schema[7.1].define(version: 2024_07_21_183116) do
   create_table "visits", force: :cascade do |t|
     t.bigint "area_id", null: false
     t.bigint "user_id", null: false
+    t.datetime "started_at", null: false
+    t.datetime "ended_at", null: false
+    t.integer "duration", null: false
+    t.string "name", null: false
+    t.integer "status", default: 0, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["area_id"], name: "index_visits_on_area_id"